requirements.txt
lxml
Pillow
Django==3.2.10
